Refactor `Command::insertWithReturningPks()` method (#324)

+++ tests/Support/Fixture/oci.sql
@@ -210,7 +210,7 @@ CREATE TABLE "default_pk" (
 
 CREATE TABLE "notauto_pk" (
   "id_1" INTEGER,
-  "id_2" INTEGER,
+  "id_2" DECIMAL(5,2),
   "type" VARCHAR2(255) NOT NULL,
   CONSTRAINT "notauto_pk_PK" PRIMARY KEY ("id_1", "id_2") ENABLE
 );
